pyproject.toml
[project]
name = "drug-deposit-dictation"
version = "0.1.0"
description = "Voice-based CLI for hospital drug inventory management"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-whisper>=20231117",
    "click>=8.1.7",
    "ollama>=0.3.0",
    "pydantic>=2.0.0",
]

[project.scripts]
drug-dictation = "drug_deposit_dictation.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/drug_deposit_dictation"]
